Added mechanism to discard older network requests (#3172)
# What this PR does

- Added mechanism to prevent concurrency conflicts when multiple network
requests are fired and the order of resolving is reversed. This resulted
in flickering the results data, thus showing discrepancy between the
filters (e.g. the search input) and the table data.

Within `RemoteFilters` and `UsersFilters` we're setting an ID for each
network result, and whenever we get to hit the store, we'll compare that
ID with the latest generated ID, thus if they match, we're all good, if
they don't, that means that network request should be discarded and its
data shouldn't be passed to the store.

This was implemented for:
- Users page
- Schedules page
(and more to come)

Fixes https://github.com/grafana/oncall-private/issues/2108