Change section title "Advanced Usage".
This section title is a bit miss leading at it doesn't offer an advanced examples and links to the basic outgoing webhook documentation.  Also, in the outgoing webhook documentation there is a section labeled Advanced Examples that leads back to the ServiceNow documentation.  This could be even more confusing for the customer.  Because of this I think we should change the name of this section to be something like "Outgoing Webhook syntax" or "Outgoing Webhook basics".  This way there isn't a circle of links marked Advanced usage/examples leading to each other.

Outgoing Webhook doc: https://grafana.com/docs/oncall/latest/outgoing-webhooks/#advanced-examples