Add alert group actions extension point to plugin (#3416)
# What this PR does

This PR adds an [extension point] to the Grafana plugin, allowing other
plugins to add menu items to the dropdown independently of this plugin.

It also moves the 'Declare incident' button into the 'Actions' menu
representing the extension point, and displays it conditionally; the
menu item is shown if all of the following are true:

1. the Grafana Incident plugin is installed (using the `PluginBridge` as
before)
2. the alert group has a valid Grafana Incident URL
3. the alert group hasn't already been linked to an Incident
4. the Grafana Incident plugin hasn't _also_ configured an identical
extension link to show in the dropdown.

Ideally the Grafana Incident plugin would soon configure an extension
using this extension point, at which point we can remove the custom
'Declare incident' button (and the `PluginBridge`) from the OnCall app.

---

The main use case I have for this is to be able to run a Sift
investigation from an OnCall alert group, similar to how it can be done
from a Grafana alert instance as of [this PR][pr].

I'll try and get the Grafana ML and Incident plugins running in the same
Grafana instance as this to test it out more thoroughly!

Note: this is heavily inspired by, and in some case stolen from, the
[Explore toolbar extension point in
Grafana](https://github.com/grafana/grafana/blob/main/public/app/features/explore/extensions/ToolbarExtensionPoint.tsx).

## Which issue(s) this PR fixes

Part of https://github.com/grafana/machine-learning/issues/3558.

## Checklist

- [ ] Unit, integration, and e2e (if applicable) tests updated
- [ ] Documentation added (or `pr:no public docs` PR label added if not
required)
- [ ] `CHANGELOG.md` updated (or `pr:no changelog` PR label added if not
required)

[extension point]:
https://grafana.com/developers/plugin-tools/ui-extensions/create-an-extension-point
[pr]: https://github.com/grafana/grafana/pull/77900