Add note about custom root CA
As discussed in issue #3898, to utilize a custom root CA, modifying the `cacert.pem` file is necessary. @mderynck  suggested a process of copying the file out of the container, making modifications, and then pasting it back inside. However, I propose a simpler approach: directly mounting the CA into the container.