fe-o11-team/upgrade-faro-and-adjust-config (#3988)
# What this PR does
* Upgrade Faro to v1.4.0
* Adjust Faro config to the new version
* Set some sane defaults

## Which issue(s) this PR fixes
* It uses `getWebInstrumentations()` to ensure that if new instruments
are added they automatically included without touching the code. This
ensures compatibility with new frontend-obervabillity features.
* Remove the old `session` object which is not available in the config
anymore.
* Enable persistent sessions
- With the new session manager Faro persists session across pageloads
and, if `persistent` session are enabled accross windows as well.
* Adding `beforeSend` hook to filter out events which do not belong to
the grafana-oncall-ap. Background: a Faro instance can not auto scope
events to a subtree of the DOM and will always capture events from the
whole window. It's advisable to filter out unrelated events.

## Checklist

- [ ] Unit, integration, and e2e (if applicable) tests updated
- [ ] Documentation added (or `pr:no public docs` PR label added if not
required)
- [ ] `CHANGELOG.md` updated (or `pr:no changelog` PR label added if not
required)

---------

Co-authored-by: Rares Mardare <rares.mardare@grafana.com>