fe-o11y-team/chore(deps): upgrade faro to v1.4.2 (#4022)
# What this PR does
Upgrade Faro to v1.4.2.

This release fixes a bug in the web-tracing package where the wrong
traceId /spanId was attached to `faro.tracing.*` events

## Which issue(s) this PR closes

Closes [issue link here]

*Note*: if you have more than one GitHub issue that this PR closes, be
sure to preface
each issue link with a [closing
keyword](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/using-keywords-in-issues-and-pull-requests#linking-a-pull-request-to-an-issue).
This ensures that the issue(s) are auto-closed once the PR has been
merged.

## Checklist

- [ ] Unit, integration, and e2e (if applicable) tests updated
- [x] Documentation added (or `pr:no public docs` PR label added if not
required)
- [x] Added the relevant release notes label (see labels prefixed w/
`release:`). These labels dictate how your PR will
    show up in the autogenerated release notes.