Fix schedule gaps when DST happens between rotation start and end date (#4103)
# What this PR does

1) IF in rotation form user selects start date before DST change 
AND according to recurrence period / selected end date it's after DST
change
AND there is a difference in hours identified because of DST change 
THEN set hour of end date to the same as start date so that there are no
gaps in the schedule

2) Fix showing current system time info in schedule slot

## Which issue(s) this PR closes

https://github.com/grafana/oncall/issues/3261

<!--
*Note*: if you have more than one GitHub issue that this PR closes, be
sure to preface
each issue link with a [closing
keyword](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/using-keywords-in-issues-and-pull-requests#linking-a-pull-request-to-an-issue).
This ensures that the issue(s) are auto-closed once the PR has been
merged.
-->

## Checklist

- [x] Unit, integration, and e2e (if applicable) tests updated
- [x] Documentation added (or `pr:no public docs` PR label added if not
required)
- [x] Added the relevant release notes label (see labels prefixed w/
`release:`). These labels dictate how your PR will
    show up in the autogenerated release notes.