Fix alert group table default filter being ignored, add 1 week window as default time range (#4080)
# What this PR does
- Default filters were being ignored on the alerts group page (Only show
acknowledged and firing). This was because the object contained an empty
array for the team field so it was never considered empty, now all
fields are checked.
- Added a 1 week window for `started_at` as a default. This should give
more useful behavior in larger instances so we won't have long queries
counting all alert groups since the beginning.

## Which issue(s) this PR closes

Closes https://github.com/grafana/oncall-private/issues/2520

<!--
*Note*: if you have more than one GitHub issue that this PR closes, be
sure to preface
each issue link with a [closing
keyword](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/using-keywords-in-issues-and-pull-requests#linking-a-pull-request-to-an-issue).
This ensures that the issue(s) are auto-closed once the PR has been
merged.
-->

## Checklist

- [x] Unit, integration, and e2e (if applicable) tests updated
- [x] Documentation added (or `pr:no public docs` PR label added if not
required)
- [x] Added the relevant release notes label (see labels prefixed w/
`release:`). These labels dictate how your PR will
    show up in the autogenerated release notes.