Feature: Use ui extension hooks where available (#4765)
**What this PR does / why we need it:**
This PR updates usage of plugin extensions APIs to take advantage of the
new hooks API where available. In older versions we fallback to the
currently used hook. This prevents an issue where due to the reactive
registry the older APIs don't receive the full list of extensions. It
also paves the way for frontend performance improvements in Grafana
core.

**Which issue(s) this PR fixes:**
Related: https://github.com/grafana/grafana-community-team/issues/174

**Special notes for your reviewer:**
We would really appreciate some assistance in testing this PR in both
the latest version of Grafana 11 and the minimum supported Grafana
version.

---------

Co-authored-by: Dominik <dominik.broj@grafana.com>