Polish unified slack ui (#4819)
# What this PR does

polish ui of unified slack:
- [x] New deprecation banner
- [x] Move install/uninstall button to the top right corner
- [x] New texts in migrate modal
- [x] New text and layout in uninstall modal.

## Which issue(s) this PR closes

Related to https://github.com/grafana/oncall-gateway/issues/299

<!--
*Note*: If you want the issue to be auto-closed once the PR is merged,
change "Related to" to "Closes" in the line above.
If you have more than one GitHub issue that this PR closes, be sure to
preface
each issue link with a [closing
keyword](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/using-keywords-in-issues-and-pull-requests#linking-a-pull-request-to-an-issue).
This ensures that the issue(s) are auto-closed once the PR has been
merged.
-->

## Checklist

- [ ] Unit, integration, and e2e (if applicable) tests updated
- [x] Documentation added (or `pr:no public docs` PR label added if not
required)
- [x] Added the relevant release notes label (see labels prefixed w/
`release:`). These labels dictate how your PR will
    show up in the autogenerated release notes.