fix(helm): Fixed migrate not being consistent in name for hooks
This makes it so that the yaml being created for the migrate job is
consistent when `useHook` is `true` or `false`. At the moment if use
hook is false, the the name on the yaml and in the metadata match. But
if use hook is true then the name on the object itself is different from
the one in the metadata.