chore: adjust dir names to irm (#4989)
# What this PR does

- rename utils to helpers
- rename types.ts to app-types.ts

## Which issue(s) this PR closes

Related to
https://raintank-corp.slack.com/archives/C0762D6EUDV/p1725477060488709
https://raintank-corp.slack.com/archives/C0762D6EUDV/p1724936143487849

TL;DR
IRM needs imported modules from oncall/incident to have uniq absolute
import paths

<!--
*Note*: If you want the issue to be auto-closed once the PR is merged,
change "Related to" to "Closes" in the line above.
If you have more than one GitHub issue that this PR closes, be sure to
preface
each issue link with a [closing
keyword](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/using-keywords-in-issues-and-pull-requests#linking-a-pull-request-to-an-issue).
This ensures that the issue(s) are auto-closed once the PR has been
merged.
-->

## Checklist

- [ ] Unit, integration, and e2e (if applicable) tests updated
- [x] Documentation added (or `pr:no public docs` PR label added if not
required)
- [x] Added the relevant release notes label (see labels prefixed w/
`release:`). These labels dictate how your PR will
    show up in the autogenerated release notes.